Update settings for platforms w/o user managemnt
Unfortunately not all platforms (cloud&heat for example) allow user
creation over the API and it's not possible to use
`cctrluser create ...` there.

Update the settings with a new flag `user_registration_enabled` and a
new property `user_registration_url` which can be used by cctrluser to
decide if it should allow the command or stop with a helpful error
message ("You can register on..").