Introduce SSH public key authentication
User will be informed that the Public Key auth is available,
letting them select it as the default way of authentication and
the path for their Public Key.
This decision is stored in the `config.json` file and can
be modified or removed from there.

The logic of this auth works this way: if the user chose
Public Key auth, we'll tell our API library to create a SSH Token.
Then, we'll put together this SSH Token, the signature of
it, the Fingerprint of the Public Key and the user email
in order to request the final User Token through the API
library, again.

If the server responses with an auth failure, we'll let the user know
and they will be asked for the password, so a Password-based Basic auth
will take place instead.