White-label home config directory
Previously, the home config directory was hardcoded
to `.cloudControl` regardless of the platform you were working on.

For instance, when creating the user authentication token,
they were all saved in this directory. It gave the possibility to
name the token file to distinguish from the default `token.json`
file.

With this commit, each platform will have a specific home config
directory. The token file is called `token.json` and it will be
stored inside the home config directory of the platform is authenticating
on.

This is a first step to introduce user specific configuration for
each platform.