Introduce setup command to handle user config
With `cctrluser setup` command a user can create or modify
their configuration whenever they need to. We do not ask
for configurations changes anymore, instead user will have
explicit control over this by using the command.

How does `cctrluser setup` work?

The command has three different flags to modify each of
the existing values on the user configuration.

`--email` will create, or overwrite if existed, the user e-mail on
their configuration. If any `--email` flag is specified,
the application will try to find the `CCTRL_USER` env variable,
if not present, it will finally ask for the e-mail interactively.

`--ssh-auth` will enable the ssh public key auth if set to `yes`,
and disable it when `no`. If the flag is not specified, value is not changed
in configuration. If this was never set (first use of `setup` command),
it defaults to `yes`.

`--ssh-key-path` specifies the path of your ssh public key used for
the authentication. If the flag is not set, it defaults to ~/.ssh/id_rsa.pub
(or its similar on Windows).

Note for exo and cnh users: ssh auth will always be disabled, regardless
of the `--ssh-auth` flag value.